Tighten up NodeNameSet#initialize signature
Injecting the MapperRegistry provides everything
needed to perform initialization. It makes the
object more self-contained as the relationship_set
to use must be the one that contains the passed in
relationship anyway. Also, we can get the needed
MapperRegistry#relation_map for free.

Also "fixes" an issue with the specs, where an
erroneous relationship mock, combined with an
erroneous relation_map mock was passing, even tho
strictly speaking the mocked relationship were
faulty (see previous let(:infos) which didn't
match the name given to the mock (:funky_infos))