Construct the minimal aliases needed for joining
This currently only works for the veritas engine.
Arel relies on Aliases#header to return an array
of Attribute::Alias instances, and the current
code produces an array of symbols instead.

This commit cleans up some other parts of the code
nicely. Aliases::Unary and Aliases::Binary are
replaced by just Aliases.

Also, see diffs for:

  Attribute::EmbeddedValue
  Mapper::Builder
  Relation::Graph::Node