Limit relation to two tuples for Relation::Mapper#one
This avoids fetching a large number of tuples when
only one is expected. Limits the relation to two
entries as this is the minimal number we need for
reporting meaningful error messages in case more
than one tuple would've been returned.

Introduces Engine::Veritas::Node#ordered which can
be considered a WIP and definitely needs more
thought.