Push Environment#reset down to TestEnv
As #reset is only used from specs currently, there
is no need to expose it as a public method to
client code. Every user of dm-mapper will interact
with Environment directly, so we should only expose
what's really necessary.

Also, the mapper registry gets populated during
Environment#finalize, so a typical user would not
need to inject an instance. It seems that this
constructor injection was primarily done to "ease"
writing of specs (using #should_receive :/). IMHO, 
until a real usecase arises, it's better to just
simplify the public api of this class and thus not
support injecting a mapper registry. It would only
be confusing for users. Also, the behavior would
be kinda "undefined" anyway.