Make dumping wrapped tuples work
I'm not exactly sure if a nested array of values is something axiom will
be happy to digest and dispatch individual writes in a joined relation.
We'll see.

The important part here is to provide more info on the attribute level
so that we know how to dump a nested tuple correctly. Its exact format
is something we can easily change later if axiom expects something else.