Expose "native" query interface through charlatan
This may look controversial as we're leaking entire interface of a
db-access library on the top level. ROM will never encourage using this
interface however for the simplicity *and* convenience sake that's how
it's gonna be. In the future we might want to hide it behind something
higher level but this would require a lot of additional layers of
complexity so I'm not sure if we'd ever want to go there.