Add basic repository feature linters
Adds a basic linter (ROM::Lint::Linter) that wraps up the common needs
out of a plain Ruby linter: a list of lints, named like test methods
(lint_*), a way to yield to each lint method, and a method to perform a
specific lint returning `true` so they are assertion friendly.

Adds specific linters for a repository and a dataset. These can be used
by repository adapter authors to verify if they comply with our api.

Updates the Minitest/TestUnit module to use the new ROM::Lint::Test
modules (TestRepository & TestEnumerableDataset) to get access to the
test method versions of the lint tests.

Many thanks to Peter Suschlik (splattael) for his help with these
features, the details of which were lost in a particularly brutal rebase
after the adapter module was moved around.