Add auto-registration setting on ROM::Repsitory
The settings:
  ROM::Repository.enable_auto_registration!
  ROM::Repository.disable_auto_registration!

Provide a way to turn off the auto-registration of repository adapters
when they inherit from ROM::Repository. There are cases, especially in
testing, where this causes problems and unexpected behavior. It is a
giant global constant afterall.

Using this will mean our specs may be more explicit but it also means we
can control the environment and test different behavior in isolation
without workarounds and complications.