Add `Command.[]` syntax to make it consistent with Relation
This makes it possible to simply inherit from a base command class
coming with the adapter or use `ROM::Commands::Create[:sql]` shortcut
syntax when registered adapters are used.

This actually simplifies things as inheritance hierarchy of commands is
simpler - no need to use modules, it's the same story as we had with
Relation.