Introduce Commands::Lazy which allows lazy evaluation of input in command graph
...in other words - this thing allows composition of multiple commands
that will process a nested input without the need to provide that input
during composition. So, you can build up a command graph for persisting
object graphs and memoize it and then simply call it with input.

This works with deeply nested input and can be mixed with commands that
return either :one or :many results with just one assumption that the
root command must be configured to return :one result. ie you persist a
user root with tasks, their tags and a user address.