Automatically build headers for relations and use lazy relations in repo
This is a nice PoC that we can easily translate a relation into a mapper
header and automatically send data through an auto-generated mapper using LoadingProxy.

Apart from a couple of small caveats like hardcoded convention for
combined keys this seems to be a good direction.

There are a couple of relation ext and smaller hacks I had to do to make
this work but it's something we can easily get rid of by improving ROM
core.