Make auto-registration use namespace value
The auto-registration code was using the namespace value as a boolean
and actually inferring the target module name from the provided
directory name. This behavior doesn't seem intuitive (if desirable it
could be set through an additional option, like `infer_namespace: true`)
and contradicts the documentation on the website, which says the
namespace option is a string.