Throw a specialized error in case a command is not found in a registry
https://github.com/rom-rb/rom/issues/324

I'm modifying the CommandRegistry initializer signature, so that it
knows on which relation name the commands are under.

I'm also using the new optional block syntax for ROM::Registry#fetch
(which works similarly to Hash#fetch), in order to throw a
specialized/descriptive error saying that it can't find a command for a
relation, as opposed to the previous generic ROM::Registry error.