Add a spec for custom dataset names with long PK names
This is a quite common convention to name PKs with a singular table
name plus '_id' suffix so it's good to explicitly support the case. The
challenge is that you have to use qualified keys on combining query
otherwise you will get 'ambiguous column name' error. The issue is
actually resolved in rom-sql.