Remove assert_tuple_count from Delete and Update commands
This served as a low-level check that we're not executing a command
against a relation that consists of more than 1 tuple *while* the
command was configured to return only 1 result.

At the end of the day this doesn't feel right. First of all we hit the
db with a count query before executing a command, which may have
horrible performance impact. Secondly, this adds *a requirement* for
every adapter that it needs to implement commands where `tuple_count` calls
`relation#count` thus expecting adapter's relation to provide `#count`.

This way we reduce the overhead for command execution *and* narrow down
the adapter interface even more.

The *downside* of this removal is that if you're using commands directly
(ie not via repository) you need to make sure that you're not calling an
update or delete command on a relation that wasn't properly restricted
(assuming it is needed).

Good news is that with rom-repository supporting commands this kind of
low-level checks are simply not needed, because command methods in
repositories restrict relations by default and expect restriction
arguments to be provided, so it's impossible to call a command by an
accident on a relation that includes more tuples than expected. ie
`user_repo.delete(1)` or `user_repo.update(2, name: "Jane")`.