Deprecate Command.validator
Validations should be handled before a command is called. For low-level
checks that can raise an exception the input processor can be used. ie
if a relation schema is defined using strict types then it'll be
guaranteed that broken data cannot be used for command execution. This,
however, should be used with caution, as *typically* data that are sent
to commands should be already validated and prepared for persistence.