Extend view DSL with support for schemas
This allows you to define a view header via schema, the block is
executed in the context of the canonical schema, so you can project it.
The view plugin will then use the projected schema to automatically
project the relation view via `Schema#project_relation`. Now, with this
in place, we can implement this interface in various adapters and there
won't be any need to manually project relations when using a relation
view. In example this will work automagically in rom-sql:

  view(:names) do
    schema { project(:name) }
    relation { order(:name, :id) }
  end

This is only a good start though, we can do a lot of cool stuff with
schemas and their attributes, ie renaming, qualifying automatically,
adding new attributes and so on. This will simplify view definitions *a lot*.