Bump to 3.0.0 (after all)
This is needed because View API has become part of the core API with a
couple of breaking changes. Going through the whole deprecation/removal
cycle wouldn't be worth the effort as upgrading is easy.

Furthermore, now all relations have schemas, regardless of the adapter,
which should also be considered as a breaking change.