Add support for combine+wrap
This makes it possible to use `wrap` and `wrap_parent` inside a combined
relation, which GREATLY speeds up loading aggregates with children
having their parents, ie posts with authors. When you use wrap operation
rom-sql will use a join with prefixed columns and automatically extract
them into a nested data structure, this results in a single query (just
like in case of combine) but the in-memory processing is trivial
compared to `combine_parents` and so it's roughly ~150-200x faster than
combine (transprocs `combine` function is just dead slow, uses
recursion, and creates a lot of intermediate arrays and hashes...we
*could* make it faster somehow, but it's still simpler and faster to
use `wrap`).

Closes #51