Always pass a relation restricted by pk to Update changeset
Previously it was broken as changeset's command wouldn't restrict the
relation. This fix also reduces coupling between update changesets and
relations as they no longer need to know about the idea of a PK, they
simply use their relation, which can be restricted however you like, and
pass it to commands. Original tuple is retrieved via `Relation#one` as
update changesets are meant to be used against a single tuple. For mass
updates plain commands should be used.