Use struct name in NoMethodError for structs with long inspection
I found that ruby has a weird behavior for representing NamerError
exceptions (NoMethodError is a subclass of NameError). When the
inspection result of an object exceeds 65 characters Ruby uses object's
class path for the error message. Not to mention it's very unexpected
and hidden behavior it does not use `.name` method defined for the class
of an object. Instead it resolves the namespace chain which is empty for
an anonymous class and that's why we get a rather unfriendly error for
struct objects which are all anonymous because of their dynamic nature.
That said I started to work on this after seeing exceptions like
NoMethodError Exception: `undefined method 'foo' for
 #<#<Class:0x007fe46e06a680>:0x007fe46c4ac5d8>`. Meh.

This PR `#method_missing` to ROM::Struct which takes the job of raising
an error in its own hands, but this comes with a couple of trade-offs.
The first is that ATM we cannot mixin the receiver into the error. I
cannot fix it without subclassing NoMethodError or patching it which is
unacceptable FWIW. Lacking of receiver leads to incompatibility with
"Did you mean ...?" and this is unfortunate. See
https://bugs.ruby-lang.org/issues/12041 for details on this. The second
is that this only fixes missing method errors on instances and does not
offer any improvement for other sorts of name errors, e.g. missing
constants, missing methods on a struct class and so on.