Add associations to memory adapter
This adds a new feature which was required to make cross-adapter
associations work, which is `override` option in association DSL. If
it's set to `true` then a given association will not produce its own
relation and will use provided `:view` instead. This way we are
completely free to define association views however we want. These views
will receive both association object and loaded relation, which should
make it more flexible.