Simplify and fix handling of struct namespace
There was a bug which caused caching of mappers even though struct
namespaced was set to a different value. While fixing it I also
simplified the way we handle this setting by adding struct_namespace to
relation AST and simply passing it to compilers. This required a tweak
in relation graph which needs to ensure that nodes use the same
namespace as the root.