Remove Command#with alias and Command#with_opts
This was an unfortunate alias since over the years we established a
convention that an object with options implements `with` method, that
returns a new instance with updated options.

Command#with was just an alias for `curry`, it's low-level, barely
anyone will be using it anyway, so it's better to make Command API
inline with the rest of the system and remove this special case when
`with` doesn't do what it does literally in every other place.

Since `with_opts` was completely redundant, and `with` is provided via
`Initializer` extension, we're left with less code. So it's a win.