Look up most nested constant first in CustomNamespace strategy
In my app I have App::Events which is a module not related to the persistence
layer and I also have App::Relations::Events which is a relations class.
CustomNamespace wrongfully picks up App::Events rather than
App::Relations::Events. This commit fixes the issue, now it starts with the
longest constant name.