[core] simplify Relation#command
This removes an awkward behavior where a command would send its results
through *all* registered custom mappers. To be honest, I don't know why
it was there. This wasn't done like that back in rom-repository, and
there are no specs nor docs that would specify this behavior.

I suspect it was an accidental left over when I was porting
auto-commands from repository to core, and programming in anger trying
to make specs pass ;)