Improve and simplify gateway config and setup
This makes gateway setup more inline with how components are handled.
It'll make it easier to extend configuration with support for global
defaults and simplify handling of gateways at runtime, because they
now have a name under which they are registered + all config and opts
associated with a gateway instance are stored within the gateway registry
which makes it easily accessible.

It didn't make sense to improve Gateway abstract class because making it
backward compatible would be too much work and breaking API would be even
more work, considering all the adapters that would have to be updated.

In summary:

  - Simplify gateway handling
  - Simplify configuring gateways
  - Add Gateway#name
  - Move config stuff from gateways environment to Configuration
  - Stop memoizing plugins in components
  - Turn Environment into a standard registry with gateways
  - Move Configuration#relation_classes to rom/compat