Unify component DSL and isolate inferring options
There's a lot of complicated logic here in order to maintain backward
compatibility. It's possible to isolate this completely in rom/compat
though, which is what I'll do before 6.0.0 release.

In general, components will be able to provide their default config
using a very simple interface, so it will be completely generic.

Basically as long as a component responds to `config` method and it
returns a Hash-like object, then default settings will be resolved
automatically. All the class attributes will be replaced by this and
moved to rom/compat.