Replace Relation's component.name setting with id and dataset
I finally realized that using relation's name object as a setting
was a mistake because it's a rich object used at runtime, it's not
part of the configuration.

Breaking it down into explicit :id and :dataset settings makes it
very clear what is what and we can easily infer them too. Still there's
additional complexity caused by backward compatibility but it will be
moved to rom/compat.

Specs are now running EVEN faster after this clean up, that's cool too.